refactor: RedisTemplate -> redisStringTemplate 로 변경(Object template 충돌 방지)
