+++ src/main/java/com/hanium/mom4u/domain/family/service/FamilyService.java
@@ -24,7 +24,7 @@ import java.util.Optional;
 @RequiredArgsConstructor
 public class FamilyService {
 
-    private final RedisTemplate<String, String> redisTemplate;
+    private final RedisTemplate<String, String> redisStringTemplate;
     private final AuthenticatedProvider authenticatedProvider;
     private final MemberRepository memberRepository;
     private final FamilyRepository familyRepository;
@@ -70,10 +70,10 @@ public class FamilyService {
         String code;
         do {
             code = generateRandomCode();
-        } while (Boolean.TRUE.equals(redisTemplate.hasKey("FAMILY_CODE:" + code)));
+        } while (Boolean.TRUE.equals(redisStringTemplate.hasKey("FAMILY_CODE:" + code)));
 
-        redisTemplate.opsForList().rightPush("FAMILY_CODE:" + code, member.getId().toString());
-        redisTemplate.expire("FAMILY_CODE:" + code, CODE_EXPIRATION);
+        redisStringTemplate.opsForList().rightPush("FAMILY_CODE:" + code, member.getId().toString());
+        redisStringTemplate.expire("FAMILY_CODE:" + code, CODE_EXPIRATION);
 
         return code;
     }
@@ -90,17 +90,17 @@ public class FamilyService {
         }
 
         String key = "FAMILY_CODE:" + code;
-        if (!Boolean.TRUE.equals(redisTemplate.hasKey(key))) {
+        if (!Boolean.TRUE.equals(redisStringTemplate.hasKey(key))) {
             throw new BusinessException(StatusCode.INVALID_FAMILY_CODE);
         }
 
-        List<String> userList = redisTemplate.opsForList().range(key, 0, -1);
+        List<String> userList = redisStringTemplate.opsForList().range(key, 0, -1);
         if (userList == null || userList.isEmpty()) {
             throw new BusinessException(StatusCode.INVALID_FAMILY_CODE);
         }
 
         if (!userList.contains(member.getId().toString())) {
-            redisTemplate.opsForList().rightPush(key, member.getId().toString());
+            redisStringTemplate.opsForList().rightPush(key, member.getId().toString());
         }
 
         // 임산부 ID로부터 가족 조회
